Separate the internal representation of path from its serialized form.
Serializing/deserializing paths from jsonArrays is certainly convenient,
but it's not the best for being able to understand how paths are
normally structured.

  - It took me quite a while to figure out what JsonNodes were allowed
in a path when. The new path has 4 explicit type of pathKeys that (I
think) are a little easier to understand
  - Metadata was implemented on path by prepending another jsonArray.
The new representation allows us to avoid that entirely, until it's time
to serialize/deserialize the path
  - Because of Metadata's special handling, it was often dropped by
JsonNodes as they iterated through path.next(). They would often grab
the next JsonNode in the path and append it directly to pathBehind
without bothering to propagate the Metadata. This new form makes it
harder to forget to propagate Metadata to pathBehind, which will be
important for assoc-in.